Simplify conservative_common_denominator
We no longer try to predict field_simplify_eq.  This results in better
behavior and less code which is more modular.  In particular, the tactic
responsible for hiding non-fraction pieces from field_simplify_eq no
longer tries to preemptively assert that denominators are nonzero.

This improvement is a result of @andres-erbsen's point in #16,
https://github.com/mit-plv/fiat-crypto/pull/16#discussion_r69035102 ,
that we were generating too many side-conditions.