A helper lemma for [Wf]
This method allows a proof term that's linear in the term being proven
well-founded, rather than exponential in it.  By factoring out all of the
reasoning about expressions, we can incur overhead equal to (the number of
constants) + (the number of let-bindings) + (the number of variables used), all
without mentioning the term itself; [vm_compute] can do the appropriate
reduction for us.