Add a non-higher-order syntax, and reg assignment
The register assigner will return [None] if you give it an invalid
assignment (too short, attempting to eliminate live code, attempting to
merge registers that can't be merged).

No correctness proofs yet, nor any sort of automatic register
assignment.  I'm planning to write a dead-code-eliminator register
assigner next.  It should also be moderately straight-forward to write a
default register allocator, or perhaps a register post-allocator, that
takes a manual register allocation, and renumbers variables in a greedy
way.  (This would let you pre-specify that some registers should be
merged, but leave the rest of the assignment up to the algorithm.)