Small example of bounds-calculation with dependent types (#61)
This might properly belong in Experiments rather than TestCase.  It demos the
ability to transform from one kind of constants to another kind, and to plug
bounds calculation functions into [var] to get bounds.  There's not currently
any sort of correctness theorem, and I'm not entirely sure what one would look
like.  Probably something that says that if you map a function over the syntax
tree and then interpret, you could instead have first interpreted and then
applied the function.

I'm hopeful that this will provide a template for integrating this version of
the syntax with rsloan-phoas.