Don't let [tauto] destruct [Equivalence] in PE
This makes progress towards #75 (now src/Experiments/Ed25519.v builds
with both 8.4 and 8.5).

Quoting [the
changelog](https://github.com/coq/coq/blob/trunk/CHANGES#L576):
> - Tactic "tauto" was exceptionally able to destruct other connectives
  than the binary connectives "and", "or", "prod", "sum", "iff". This
  non-uniform behavior has been fixed (bug #2680) and tauto is
  slightly weaker (possible source of incompatibilities). On the
  opposite side, new tactic "dtauto" is able to destruct any
  record-like inductive types, superseding the old version of "tauto".

There was an `Equivalence` in the context that `tauto` was destructing,
but otherwise not using.  This made the proof pick up an extra
dependency from the context in 8.4, but not 8.5/8.6.