Clean up fe25519_word64ize, give examples
As per
https://github.com/mit-plv/fiat-crypto/commit/dc8a141d13984bc39cf7523856b678a48428dbcc#commitcomment-19672297

Previously, the proofs were not unfolding the boolean comparisons that
eliminate opaque proofs from the resulting proof term.  We have to jump
through a number of hoops because Coq is really bad at dealing with big
terms, and so since we can't unfold all of the identifiers early, we
unfold some of them and alias other ones, so that we can unfold the
right ones with a hopefully-not-too-long cbv list at the end.  (This
would be a lot prettier if wish #4473
(https://coq.inria.fr/bugs/show_bug.cgi?id=4473) were granted.)