Generalize add_coordinates
This is in preparation for dropping extra carries.  What remains to be
done, after this and #106, is to finish packaging up the reified
[add_coordinates] so that it can operate on [Tuple.tuple
GF25519BoundedCommon.fe25519 4], and then to prove

```coq
forall twice_d P1 P2, Tuple.fieldwise GF25519BoundedCommon.eq
  (<reflected add_coordinates> twice_d P1 P2)
  (@ExtendedCoordinates.Extended.add_coordinates GF25519BoundedCommon.fe25519
        GF25519Bounded.add GF25519Bounded.sub GF25519Bounded.mul twice_d
   P1 P2)
```

I'm not sure how to do this, or even what the right structure for the
proof is.