Update _CoqProject
Version-controlled .v files should not be removed from _CoqProject.  We
can filter them out in the Makefile as necessary so that they don't get
built by default, but if we desynchronize _CoqProject, then we can end
up with confusing error messages that depend on which .v.d files are
left over from previous builds.