CSE without inlining arithmetic expressions
This takes care of most of #158.  The remaining bits are reworking the
Wf and interpretation lemmas to actually work.  (The former needs a only
bit of rethinking and rephrasing to handle the fact that sometimes we
change the stored symbolic expression from a complicated one to a fresh
variable, while the latter needs major surgery, which Adam tells me is
easy, and this is a note that when I come back to it, I should look at
the email thread with Adam about CSE from last summer.)