Factor karatsuba through IdfunWithAlt, add test
Currently the refinement is commented out.

Also, we drop the proof of equality early (and probably won't use it in
the first place); there's no way we can carry around such a proof in
reflective-land, so we'll need to add an arithmetical-equality
semi-decider to reflective-land that can prove the relevant equalities
(or we'll need to leave them over as side-conditions).  I expeect this
may make things significantly easier on @jadephilipoom.