More WIP on register allocation
The current allocation is terrible, probably because we are currently
requiring that all instructions output to registers.  My current guess
at a decent thing to do is to make a pass, after register allocation,
and eliminate all registers that simply get stored to memory,
replacing the relevant instructions with the memory-using versions.
Then we can re-register allocate, ignoring values that go straight to
memory.