Move chained_carries' (now chained_carries_reduce)
It now lives in Arithmetic.Core.B.Positional, where it belongs, rather
than in Specific/.../HelperTactics.

Andres notes that we probably don't need this at all, and could instead
make chained_carries reduce after every index (and the spurious
reductions should be no-ops).  I didn't want to bother verifying this,
at the moment, so I left it as-is.