Update make_curve.py for python3
In python2, we require unicode.  In python3, "unicode" as a type doesn't
exist.  In most places, we can replace the isinstance check for
str/unicode with basestring.  In one place where we need to output
unicode, we do a terrible hack to set unicode to str if it doesn't
exist.