benchmark montgomery32, gmpvar, and gmpsec on 32-bit android
Despite significant effort, I did not manage to compiler gmpxx in a way
that would allow it to be dynamically linked with the c++ standard
library on android, or statically link the library in a way that would
result in a file executable on Android. Somebody who understands C and
Android dynamic linking might be able to do better.