Partial work on fixing partial reduction
I've mostly fused partial_reduce' with arguments.expr.interp.  This
fixes some things, but it doesn't fix everything.  In particular, the
algorithm is set up incorrectly to run in a single pass.  I think the
correct fusing treats the [arguments] metadata as something like
types-in-bidirectional-typechecking.  That is, we need to drive the
evalutation by expected [arguments], except at operation nodes, where
we feed the expected output [arguments] to the rewriter and it tells us
what sort of [arguments] we need to use to drive the partial evaluation
of the op-arguments.

Arrows are kind-of wonky, though.  What's the right argument annotation
for something like [@List.map A B]?  The simplest thing is [λ '(list b),
(A -> b) * list A], i.e., if you want it to produce [list b] for some
description [b], you need to give it a function from a generic
expression for [A] to a thing with as much info as is in [b], and you
need to give it a list of generic [A]s.  But that's not accurate,
because it actually works fine for any consistent detail on [A].  So we
want something like [λ '(list b), ∃ a, (a -> b) * list a], but I don't
know how to express that.  This is where we do want something like
bidirectional typechecking: we drive the evaluation of [List.map f x] by
the description [list b] of the output.  Then we drive the evaluation of
[f] with the description [?a -> b], and we want to get out [?a].  Not
sure how this all fits together though....