[experiments] Use smaller multiplication for 19*x (#307)
See also #300.

We take the simple path of pulling all of the constant multiplication of
sufficiently small size to the end of any expression like a * b * c * d
* ..., and then we associate all addition and multiplication all the way
to the right.  This might be very much the wrong choice in some cases;
I invite suggestions that don't require knowing any bounds, which
perform better.  (Doing things that require bounds would complicate this
significantly.)

I have no idea why this reassociation results in this part of the diff:
```diff
  expr_let 15 := (uint64)(x_14 >> 51) in
  expr_let 16 := ((uint64)x_14 & 2251799813685247) in
- expr_let 17 := x_4 +₆₄₋₆₄₋₆₄ (19) *₃₂₋₆₄₋₆₄ x_15 in
- expr_let 18 := (uint32)(x_17 >> 51) in
+ expr_let 17 := x_4 +₆₄ x_15 *₆₄₋₆₄₋₆₄ (19) in
+ expr_let 18 := (uint64)(x_17 >> 51) in
  expr_let 19 := ((uint64)x_17 & 2251799813685247) in
- expr_let 20 := x_18 +₃₂₋₆₄₋₆₄ x_7 in
- expr_let 21 := (uint32)(x_20 >> 51) in
+ expr_let 20 := x_18 +₆₄ x_7 in
+ expr_let 21 := (uint64)(x_20 >> 51) in
```