Add a constant-time binary extended GCD algorithm.
Thankfully, most modular inverses involve a prime or at least odd
modulus, but RSA key generation requires inverting with an even modulus
(the totient). Implementing that in constant-time is tricky.

This file adapts the binary extended GCD algorithm in the Handbook of
Applied Cryptography. This is to verify the C implementation in
https://boringssl-review.googlesource.com/c/boringssl/+/26370.