Move cps out of cache part of the pipeline
Most of the changes are due to the fact that the PrePipeline (cache)
part no longer errors, and the pre-check post-cache part of the pipeline
can now error.

Note that this slows rcarry_mul down from about 0.3 s to about 2.6 s
(some of which is probably counter-balanced by the fact that caching is
probably now faster).