Don't introduce extra lambdas and apps in uncurry
Previously I was trying to make the proof easier by using the same var
type for input and output (which would allow a
correctness-of-interpretation proof which doesn't depend on
well-formedness).  We now no longer do that, and instead go from `@expr
(@expr var)` to `@expr var`, and avoid introducing useless `Abs` and
`App` nodes.