Pass around lists of strings for error messages
This allows us to get back an error message in Saturated Solinas in
OCaml on P256 x32 in reasonable time without blowing the stack.

There's the slight oddity that the list of string in the success case is
joined by the empty string, but the list of string in the error case is
joined by newline.  Probably meanas that I chose the wrong abstraction
barrier somewhere.