Add interp-correctness condition for rewriter
The NBE-rewrite rules are proven correct.  The arith and fancy rewrite
rules are reduced to lemmas about Z, most of which are inconsistent (and
therefore indicate rewrite rules which are missing side-conditions).

One bad rewrite rule was found in this process, and corrected in
0842563b23f8d780f4ff1080d7620fc3f368191f

The next step after this will be using the rewrite rule correctness to
prove the rewriter correct.