Guarantee that casting always returns inrange
We fuse the bounds relaxation pass with the abstract interpretation pass
by folding bounds relaxation into the annotation method.  This allows us
to not need any particular assumptions about what happens when you cast
outside the range.

We can therefore now guarantee that the output of casting is always
bounded by the given range (rather, by the normalized version of the
range), which will be required to take advantage of bounds information
in the rewriter (because otherwise we're not guaranteed that casting
something to within a range actually outputs something within that
range).