Add placeholder rewrite rules for rewriting after bounds
Currently, there aren't any interesting rewrite rules in the lists, but
this allows us to move rewrite rules that depend on cast behavior to
after bounds analysis.

Unfortunately, it takes a lot of time to build from Rewriter.v to the
standalone binaries, so it'll probably make sense to inline some stuff
for more rapid development.

Note that we also now have a tactic that does all of the evaluation for
rewrite rules (assuming the right [Arguments] commands have been
issued).