Restrict [ident.cast] a bit more
We now enforce that [ident.cast] commutes appropriately with [Z.opp] /
[ZRange.opp].  In particular, we perform normalization to ensure that
more of the range is positive than negative (and if we have a range like
`r[-x ~> x]`, then we ensure that the value being casted is
non-negative), before using the given cast-outside-of-range function.
Afterwards, we perform the relevant negation to set it back.

This makes a number of rewrite rules simpler and easier to prove,
because we won't have to ensure that the casts are the same pre- and
post- negation; it's guaranteed that we can mess around with negation as
much as we want and still have things work out fine.