Be stricter about rejecting command line arguments
We no longer accept `1,1` as a valid bitwidth.

This closes #488.  It does not quite fix the issue as stated, but it
resolves the underlying confusion that generated it (the difference
between passing `1,1` as a `c` vs passing `1,1` as a bitwidth).  It
still might be nice to someday emit warnings as suggested in the bug
report.

I've also factored out a bit of code to make it easier to report parse
failures, by folding over a list of parses.