Add support for splitting multiplication
This adds a rewriting pass (disabled by default) which can split apart
multiplications to emit code in a single integer type.

Note that we can emit PHOAS code with only a single integer type, but to
emit C code, we must also allow uint1; see
https://github.com/mit-plv/fiat-crypto/issues/524#issuecomment-529060374
for more details.

We leave it like this because the bedrock2 translation will be coming
from PHOAS.

Note that we must still permit wider integer types during bounds
analysis, because otherwise we'll fail to assign bounds before we even
get to the point of splitting multipliciations.

Note that this rewriting pass is still admitted, so this commit adds a
dependency on an axiom to the overall pass.  (We could create a separate
pipeline if we wanted the code we're extracting to not depend on this
admitted proof.)

The interesting files to look at here are `src/RewriterRules.v`
(defining the rewriting rules) and `src/SlowPrimeSynthesisExamples.v`
giving examples.  The admitted proof is in `src/RewriterRulesProofs.v`.