Add CPS'd versions of arithmetic operations
This is primarily for perf testing purposes, so we can evaluate the time
spent in the rewriter vs cbv/vm/native/etc for the cps'd versions.

We evaluate vm_compute and native_compute by parameterizing everything
over a module of runtime definitions, which can be instantiated with
module-locked opaque constants, which are not unfolded by the vm.