Refactor the CLI to make modifications easier
There is now a Class which defines what the pipeline-specific parts of
the CLI are (pipeline-specific argument parsing, usage/help string,
display of arguments, synthesize, check_args), so that we can define
these separately for each pipeline, but not have to copy most of the
boilerplate.

It's plausible that argument parsing, help strings, etc, should be
factored through Crypto.Util.Arg, which implements roughly the API for
OCaml's Arg module for command line parsing.  This seems like a very
heavy hammer to use, though, so for now I'm leaving the argument parsing
as specialized (and therefore slightly duplicated).

This will hopefully make it a bit easier to only have to add
output-language-command-line logic in one place.

Alternative to #559

Closes #559