Move a template inductive outside a section
For compatibility with coq/coq#9918

Closes #564

We could instead have just removed all of the Compilers/InSet stuff, as
nothing depends on it (and the issue it was meant to work around in
reification by parametericity is (a) fixed by using `type of` in Ltac
and (b) doesn't even seem to be used in the reification pass).

Instead, I chose to leave the existing code mostly as-is, and insead
move the one inductive that needs to be template-polymorphic out of a
section.  I chose this because it seems to be the decision that is
marginally in the direction of stress-testing Coq more.