Hard-code overflow behavior to be wrapping
This is required for dropping the carry in the mul-split addition
rewrite rule.

We also insert a compile-time test into the C code for twos-complement
wrapping behavior.

Question/TODO: Should we insert a test or use, e.g., `Wrapping`, in the
Rust code?

Note that this weakens the guarantee we get from bounds analysis; now
our guarantee is that we perform correctly when doing twos-complement
wrapping on overflow, rather than that we perform correctly regardless
of overflow behavior (as long as it is well-defined).

Note also that we leave everything parameterized over the now-useless
`cast_outside_of_range` so that we don't have to change large amounts of
code just to support targeting bedrock2.