Relax bounds on literals during stringification
This results in fewer useless casts on literals, and fixes the
`to_bytes` Go synthesis issue.

The underlying assumption required by this change is that integer
literals don't need to be casted to be recognized as the narrowest
bitwidth that they fit into of the bitwidths we allow.  (This is
trivially true in C, at least.)