Update the multiret rewrite rules
They now handle constants in adc/sbb better.  We also split the lists
and reindent a chunk of the arith_with_casts rules, in preparation for
adding a new optimization.  Doing this here will make the new
optimization more obvious.  Note that the change here doesn't actually
change any of the rules used in the C files.