Give stringification access to machine_wordsize
This will make bedrock2 32-bit support significantly easier.

Currently the CLI support is a bit messy, because I couldn't decide
whether bitwidth was "owned" by the overall pipeline (in which case it
probably shouldn't show up in the header, or should show up uniformly
with all the other options in the header, like static and some rewriter
options), or whether it's "owned" by individual synthesis pipelines
(like m, s, and c) (which allows it to continue showing up as its own
line in the header, as it currently does).

Probably we should eventually choose the former, but because I haven't
really thought out this design decision at this point, I currently
preserve the way the output files are currently being generated and do
this sort-of mixed support mode.