Synthesize primitives witout the bytes size
As per
https://github.com/mit-plv/fiat-crypto/issues/706#issuecomment-657154861

We've been synthesizing `cmovznz` and telling it that the `bytes` size
is allowed at
https://github.com/mit-plv/fiat-crypto/blob/e5751ea93dd729d55c83eb0896d7f321975bcd2c/src/PushButtonSynthesis/Primitives.v#L842-L850,
but when we synthesize, e.g., `opp` at
https://github.com/mit-plv/fiat-crypto/blob/e5751ea93dd729d55c83eb0896d7f321975bcd2c/src/PushButtonSynthesis/WordByWordMontgomery.v#L347-L355
we tell it that the `bytes` size is not allowed.  This hasn't caused
issues before, because I don't think I have any other examples where
carries should be widened to something wider than bytes.