Don't widen carries in Go, define wrappers
Unfortunately, there's not really a good way currently to widen carries
to bytes, as some parts of the codebase see the bytes/no-bytes as a
function-local destinction, and there's no way to widen carries to bytes
while widening things wider than carries to bitwidth.  See also
https://github.com/mit-plv/fiat-crypto/issues/846

So instead, for Go, we don't widen either, and define types for carries.

Additionally, we wrap bits.{Add,Sub}{32,64} so that the outputs and
inputs have the right sizes.