Generalized sub/opp balance for unsaturated solinas
Alternative to #857.

Rather than redistributing an existing encoding, this version simply
encodes a number using `partition` that, when added to the relevant
bounds on a limb-by-limb basis without carrying, gives something which
evaluates to the desired result.

The limb-by-limb addition ensures the bounds we want, and the encoding
scheme ensures the correct value.  Furthermore, we don't need to know
ahead-of-time how many copies of the prime we need to add. This is
pleasingly simple to me.

Additionally, as in #857, the tweak to make `to_bytes` work for p224
(and other primes with low limbs) and unsaturated solinas.  The tweak
here is a bit uglier than in #857, because balancing here doesn't
guarantee the same value under `eval`, only something congruent to that
value.  So we need to ensure that we're encoding at least one copy of
the prime.