Use the redistributive algorithm for balance
As per the consenus of the group meeting.  Unfortunately, this algorithm
computes incorrect balance for some primes (leaving over negative
limbs), as seen in the changes to the test-suite.

If the algorithm worked as expected, I believe the two algorithms would
give exactly the same answer in all cases where their evaluation is
equal and the input does not exceed the maximum of the weight and
lower-bound in any limb except possibly the highest one.  (This last
condition morally says that we have not already rebalanced for a larger
minimum value.)  Here is my argument:

The specification of the algorithm of #857 (henceforth algorithm (1))
is that it computes the minimum value that each limb can have subject to
the following constraints:
1. the overall evaluation must equal that of the input
2. each limb must be greater than or equal to the specified lower bound
3. each limb remains unchanged iff, after borrowing what's necessary for
   the lower limbs from the current limb, the current limb is already
   above the minimum value.

The specification of the algorithm of #860 (henceforth algorithm (2))
is that it computes limbwise the number:
```
partition (eval (input - minvalues) mod p) + minvalues
```

The intuition is that both algorithms in fact compute the minimum
possible borrow for each limb.

The argument:
Note that if both algorithms give results that evaluate to the same
value, then both algorithms can be expressed as borrowing some amount
from each non-lowest-limb to add into the next-lower-limb, by the
properties of modular arithmetic (in particular that `x = x / d + x mod
d` for every `x` and any positive `d`).

Then the two algorithms compute different limbs iff they borrow
different amounts from some limb to add to the next-lower-limb.
Suppose, for contradiction, that there is a such a limb.  Say that the
lowest such limb has index `i+1`, and that (1) borrows `b1` from limb
`i+1` for limb `i`, while (2) borrows `b2` from limb `i+1` to add to
limb `i`.

Let us compute the output limbs of (2).

By the specification of `partition`, limb `i` of the output of (2) is
equal to
```
minvalues[i] + (eval(input - minvalues) mod p) mod (weight(i+1)) / weight i
```
The assumption that the evaluations of (1) and (2) are equal is the
assumption that `eval(input - minvalues)` is already reduced modulo `p`.
Hence we have that limb `i` of the output of (2) is equal to
```
minvalues[i] + eval(input - minvalues) mod (weight(i+1)) / weight i
```

Since `(x mod weight(i+1)) / weight i` is less than `weight(i+1) /
weight i`, we have that the amount we add to `minvalues[i]` is at most 1
more than we would need to borrow to get exactly `minvalues[i]`.
Furthermore, if we can get exactly `minvalues[i]` by borrowing some
amount, then we borrow exactly that amount, and not more, because the
remainder will be smaller than `weight i` in that case.

Hence if we borrowed any less than `b2`, then we'd end up with something
less than `minvalues[i]`.

By the constraints on (1), `b1 ≤ b2`.

Hence `b1 = b2`, in violation of our assumption that they differ in this
limb.

Qed