Play some tricks to remove axioms from coqchk
We provide dummy definitions locked inside opaque modules rather than
using axioms for constants to be extracted as stateful primitives.

Since all we care about with coqchk is that we cannot prove False, it
does not really matter if the model we provide for these constants does
not match their semantics. We only use them for interaction with the
system in any case.