Remove dependencies on the real numbers
By copying over the implementation of nsatz, we can drop the
dependencies on the real axioms.  Together with
https://github.com/coq/coq/issues/12845, this should eliminate the last
axioms reported by coqchk, allowing us to be done with #736 (maybe
modulo adding a CI job to check that no axioms creep back in).

Not clear to me if this is worth it...