Remove v8.8 and v8.7 from CI
For thoroughly unclear reasons, the tip of these branches results in
notations printing incorrectly, while the latest released versions print
fine.  Furthermore, this seems to only happen on the CI, and not on the
my locally built version of the tips of these branches.  So I'll just
remove the tip versions from CI :-/