Pipe through an always-false equivalence checker
This should be enough rope to check equivalence of PHOASTs with assembly
expressions.  It's plausible we want more control over the calling
conventions from the command line, for example, specifying:
- what to do when we run out of registers / how arguments are stored on
  the stack
- support for a number of common calling conventions

We may also want to emit better headers for the assembly files, rather
than replicating the header of the input file.