Eliminate addcarryx/subborrowx wrappers in Go
This is a partial fix for item 1 of
https://github.com/mit-plv/fiat-crypto/issues/949#issue-870337237 also
mentioned at https://github.com/golang/go/issues/40171#issuecomment-890987766

Unfortunately, we still cast the arguments to uint1 because we don't
have enough information in the right places to relax the casts.  We
really want a fused absint/rewriting pass, where we get access to true
bounds information rather than just casts.