[go] Widen the carry type to 64 bits artifically
For performance reasons.  Note that this is unsound, but having the
carry type be 8 bits rather than 1 bit was already unsound, and this is
no more unsound.  The narrowness of the carry type is not used in
practice in the code, though we should still someday include it in the
proofs.