Add `--no-shiftr-avoid-uint1`
We now default to avoiding uint1 in the result of `>>` operations, as
per
https://github.com/mit-plv/fiat-crypto/pull/1015#discussion_r708691741.

This is required for making symbolic equivalence checking with assembly
go through.  It might be possible to improve the symbolic equivalence
checker to know enough about bounds analysis (or to also do bounds
analysis) so that this is not necessary?