More assembly equivalence WIP
Only two subgoals remain:
```coq
  R_runtime_output frame runtime_rets
    (type_spec_of_runtime (word_args_to_Z_args word_runtime_inputs))
    (Datatypes.length x) stack_base asm_args_out asm_args_in callee_saved_registers
    runtime_callee_saved_registers m'
```
and
```coq
 R
   (frame ⋆ R_list_scalar_or_array x0 asm_args_out
    ⋆ R_list_scalar_or_array (word_args_to_Z_args word_runtime_inputs)
        asm_args_in ⋆ array cell64 (word.of_Z 8) stack_base x)%sep x1
   (update_dag_with (init_symbolic_state d) (fun _ : dag => d0)) m
 ->
 R_mem frame G_final dag_state1
   (rev (List.combine x4 x5) ++
    rev
      (flat_map
         (fun pat : (idx + idx * list idx) * (idx + list idx) =>
          match pat with
          | (inl _, _) => []
          | (inr (_, addrs'), inl _) => []
          | (inr (_, addrs'), inr items) => List.combine addrs' items
          end) (List.combine x3 (l4 ++ inputs))) ++ init_symbolic_state d) m
```

The first goal is that we must connect everything from the final
symbolic evaluation with our output spec.

The second goal is that we must connect the memory predicate from before
SymexLines (post-initialization) with the desired memory predicate, by
showing that we've updated symbolic memory as desired.

Note that we'll probably want some `iff` statements around `R_mem`,
because what we do in one direction in the first goal we do in the other
direction in the second goal.