Parameterize over dereference_scalar in more places
This will allow us to reuse the same code to load inputs and outputs,
when we want to enforce the spec of reading back and removing all of the
values and then saying that the resulting symbolic memory state is
empty.

Some of the proofs are not quite as complete as they were before, but I
hope to replace them soon with a more principled approach.