Add {,Bedrock/}Everything.v that requires all files
This will check for universe inconsistencies and notation clashes.  This
should fix a current issue where we catch universe errors in coqchk
only, which only runs on master and not in PRs.

The file is automatically regenerated before being built.

We have two versions, one for building with bedrock and one for building
without.