Add `with_bedrock2` binaries (#1252)
This is in preparation for #1250.  We generate `with_bedrock2_*`
binaries which support bedrock2 output but default to C.  We also
install these by default when we are installing with bedrock2 enabled.
These will be the source files that we want to package up into source
releases of our generated code, as well as the source we use for
generating statically-linked redistributable binaries.