Friendlier GH Actions Queue (#1280)
We now cancel previous queued and in progress actions runs on PRs when a
PR is updated. Because [`github.head_ref` is only defined for PRs and
`github.run_id` is
unique-per-workflow](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-using-a-fallback-value-1),
we won't cancel previous jobs on branches, meaning that we'll still see
a log of which commits succeeded and which ones failed.