Reorganize dag modules to enforce better abstraction (#1508)
We still only enforce it in Symbolic.v, but hopefully that'll be enough
to mostly ensure that future changes to one dag representation don't
desync with the other one.

Hopefully at some point I or someone else will manage to track down the
performance issues in switching to the New dag representation, and then
maybe we'll be able to drop this?  Or it might be good to make it easy
to change dag representations in the future.

In doing this, I've taken the opportunity to deduplicate some proofs
between the new and old implementations.