In its state before this commit, the reduce_carry_borrow function did rather unreasonable things when the limbwidth was fractional.
In this commit, I tried to fix this issue, but I ended up making the
reduce_carry_borrow function pretty obnoxious.  Also, fixing this issue
requires somewhat strange (and strong) hypotheses on the weight function.
We'd want, for instance,

(weight (2 * limbs - 2) / weight (limbs - 2)) mod s = 0.

I plan to revert my changes (from this commit) to the reduce_carry_borrow
function, so that it will continue to do unreasonable things given fractional
limbwidths.  This was not worth it.